feat: move stdout client to a separate package
